[FEAT] 2번까지 Make 성공 (#1)

+++ project3/proc.h
@@ -60,7 +60,7 @@ struct proc
 	struct file *ofile[NOFILE]; // Open files
 	struct inode *cwd;			// Current directory
 	char name[16];				// Process name (debugging)
-	int priority;				// Process priority
+	long long priority;				// Process priority
 	int weight;					// Process weight
 };
